+++ apps/remnant2toolkit/app/(items)/item-lookup-v2/_components/submit.tsx
@@ -1,0 +1,13 @@
+'use client';
+
+import { BaseButton } from '@repo/ui';
+import { useFormStatus } from 'react-dom';
+
+export function Submit() {
+  const { pending } = useFormStatus();
+  return (
+    <BaseButton type="submit" disabled={pending}>
+      {pending ? 'Submitting...' : 'Submit'}
+    </BaseButton>
+  );
+}